refactor : all

+++ Assets/Wall.cs
@@ -2,7 +2,7 @@ namespace KSY.Tile
 {
     public class Wall : TileObject
     {
-        public override void Magnetization(sbyte xDir, sbyte yDir, TileObject presser)
+        public override void Magnetization(int xDir, int yDir, TileObject presser)
         {
 
         }
